requirements.txt
flask==3.1
requests==2.32
beautifulsoup4==4.13
python-dotenv==1.1
